chore: simplify the netlify deploys GitHub Workflow
No more gh artifact download/upload.
We don't need to store it as we can download the published site as a zip from
the netlify dashboard.

No more need to run on 'pull_request'.
We don't need the full site as we have a preview for PR. The preview is managed
by a dedicated workflow which can be temporary tuned for specific needs.