feat(redirect): Add redirect for next dev version (#262)
`/next` will redirect to `/1` (aka the alpha version) without changing the URL. 
The main url is - and will remain - `/1`, because it allows us to ensure that the dev version is sorted correctly among other versions (at the end). So a user that go to the alpha version through the menu will have the URL `/1`. But at least, with this redirect, if we want to send a link to a client we can do it with `/next`, which is a bit better.