test: allow use of project document content and default ui bundle
Provide document content directly in the project for testing purpose. The site
is fast to build with this content.
It lets test UI bundle change, Antora upgrade, site structure and much more.
It is also possible to build the site with the Antora Default UI for comparison
with the Bonita Theme.