chore: bump Antora from 2.3.4 to 3.0.1
This bump doesn't require any change in the documentation content and only minor
changes in the Antora configuration.

Introduce a custom Antora extension to manage versions order
We want alpha prerelease to be at the end of the version list, we previously
relied on the '1' version, as all versions were sorted by version key. This has
changed in Antora 3: prerelease versions are at the beginning of the list.
The custom 'versions sorter' Antora extension ensures that all the alpha prerelease
are put at the end of the list.
This is the recommended way to manage such a use case (see discussions with the
Antora community in zulipchat during the alpha/beta tests).
Notice that for bonita, the alpha prerelease was previously put before the
'archives' version. It is now put at the very last position.

Better error detection (in particular broken xref)
Always use the 'pretty' log format to have readable log on CI
Only enable 'sourcemap' for preview (perf impacts). This provides more info
about the file location (line number) where errors and warnings occur.

Update the documentation content used for tests
  - remove broken xref in the cloud home page (detected thanks to Antora 3 logs)
  - add asciidoctor 2 syntax in bonita home (in the latest version) to demonstrate
  the new 'collapsed block' feature