chore: move special CI configuration out of the preview script (#446)
Historically, the preview script was directly called by a lot of GitHub
workflows. To avoid to duplicate the `npm ci` setup and ensure that
Antora fetches sources, it was easier to do it in the script.
Now, all documentation content repositories use the GitHub actions
provided by the bonita-documentation-site repository, so the extra
settings can be managed by the shared actions.

This move provides a better separation of concerns. The script only
focuses on generating the preview without managing special cases
depending on the environment. The `npm ci` command is called directly in
the `build-setup` internal action, where everybody expects to find it.
This allows to introduce the usage of the GitHub cache for the npm
dependencies, so we can expect fastest builds.
This also ensures consistency between the build and publish PR preview,
and the "publishing to production" workflows that are no more doing `npm
ci` in a different way.

In addition, we now run and advice to run Antora via `npx` as suggested
by the Antora documentation.