ci(preview-action): add ignore-errors input (#466)
The new input allows to build and deploy the PR preview but also to
validate xref.
Currently, this is needed for documentation component that have no xref
to other components. See
https://github.com/bonitasoft/bonita-labs-doc/pull/139
In the future, when `Antora Atlas` will be used (#326), all
documentation content repositories will progressively do xref validation
as part of the preview build.