refactor: use the "version-alias" AsciiDoc page attribute
The Antora model now only contains the version alias value to add to the search parameters if it is
present.
This reduces the logic in the UI, avoids to hard code the latest/next alias in the UI.
Also improve the extension logs and rename the extension to match the functional changes.