refactor(preview-script): fix typo in `ignore-errors`
Use plural for consistency with other options.
This option is not supposed to be used outside of this repository. The shared GitHub Actions provide
a dedicated input which must be used instead of the script option.
So, this change won't have side effects in workflow of the documentation content repositories.