ci: make PR surge deployments work with both site and test (#713)
This was not working as the 2 workflow used to deployed called the same
reusable workflow that deployed to the same URL for the 2 environments.
This was because the URL is computed using the job id of the reusable
workflow.
So for "test" deployment, duplicate the logic of the reusable workflow
to have a dedicated URL.

In the "deploy" reusable workflow, use a shorter job id. It is used to
compute the URL of all deployments. A shorter name ensure that we don't
try to generate a too large URL subdomain.

The PR comment relative to the changes will be managed by a dedicated
workflow in the future for a better separation of concerns, to simplify
the maintenance and to make things more explicit. So remove the related
steps (that was still a work in progress) from the reusable workflow in
charge of the deployment.

Covers #686