ci(surge-deploy): get artifact from the specific workflow run
Previously, this wasn't set in `dawidd6/action-download-artifact` so the
action was doing a search to retrieve the artifact.
It was using default search parameters, so it skip workflow runs for PR
from forks, and always considered the latest run of the workflow.
So, when several PR was opened simultaneously, a wrong artifact could be
retrieved.

In addition, don't run the deployment workflow if the build step failed.
Fix the fix, the deployment would fail because no artifact is available for download.