ci: move internal actions to reusable workflows
The "build-and-publish-pr-preview" and the "build-pr-site" actions are no longer used in any
bonitasoft repository, some new reusable workflows are used instead.

This makes possible to remove the "build-and-publish-pr-preview" action and to move
the "build-pr-site" and the "upload-pr-built-site-artifact" actions code directly to the "validate
references" workflow. The old actions have then been removed.

For consistency with other repositories, the workflow that validate the references in this
repository has been renamed.