ci: rename job id of "_reusable_surge-build-preview.yml" (#820)
The current job id is used by the surge-preview to build the deployment URL.
This URL is used to build the site and to check its validity.

However, it was not the same as the actual one used for deployment as the job id was not the same as in the workflow doing the deployment.
In addition, the former job id was longer so in repository with a long name (like for BCD), the URL was detected as invalid (because it was too long).

The workflow now uses the same job id as in other workflow, which fix all issues.