ci: add a GH workflow to detect broken external links
The link check uses `htmltest`: https://github.com/wjdp/htmltest and a dedicated site built using the preview script (which has been updated to support this new use case).
Reasons for using a dedicated site:
- only build the supported version: out-of-support content is not updated
- no "edit this page": generate a lot of links to GitHub, this is an Antora feature, we don't want to test that
- force usage of the production navbar to have the content as close as possible of the production site
- use a dedicated "type" option to produce convenient HTML files for URL checks (HTML extension for local file browsing, no redirect pages)

The site is checked locally by browsing its pages. Not relying the network to access to the page fasten the execution.

`htmltest` provides a cache mechanism to limit the check of URL recently checked.
As the execution is fast without this cache, it is not used for now which simplifies the maintenance.
Otherwise, we would have to setup a mechanism to store and update the cache on regular basis).

The README has been updated to explain how the links check is done and how to perform the checks locally.