ci: simplify the reusable workflow that builds the surge preview (#840)
There is no need to run this workflow when the PR is closed. When the PR is closed, the site is left as it is or it is undeployed.
So, there is no need to build the site, and the whole workflow execution can then be skipped in this case.